+++ nostr/nostr-base/src/test/java/org/tbk/nostr/base/EventUriTest.java
@@ -10,30 +10,30 @@ class EventUriTest {
 
     @Test
     void testEquals0() {
-        assertThat(EventUri.fromString("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2"),
+        assertThat(EventUri.parse("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2"),
                 is(EventUri.of(1, "82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2")));
 
-        assertThat(EventUri.fromString("2:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:"),
+        assertThat(EventUri.parse("2:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:"),
                 is(EventUri.of(2, "82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2", "")));
 
-        assertThat(EventUri.fromString("3:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:test"),
+        assertThat(EventUri.parse("3:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:test"),
                 is(EventUri.of(3, "82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2", "test")));
 
-        assertThat(EventUri.fromString("4:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2: : test :"),
+        assertThat(EventUri.parse("4:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2: : test :"),
                 is(EventUri.of(4, "82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2", " : test :")));
     }
 
     @Test
     void createFromString0() {
-        assertThat(EventUri.fromString("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2"), is(notNullValue()));
-        assertThat(EventUri.fromString("2:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:"), is(notNullValue()));
-        assertThat(EventUri.fromString("3:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:test"), is(notNullValue()));
-        assertThat(EventUri.fromString("4:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2: : test :"), is(notNullValue()));
+        assertThat(EventUri.parse("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2"), is(notNullValue()));
+        assertThat(EventUri.parse("2:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:"), is(notNullValue()));
+        assertThat(EventUri.parse("3:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:test"), is(notNullValue()));
+        assertThat(EventUri.parse("4:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2: : test :"), is(notNullValue()));
     }
 
     @Test
     void createFromString1() {
-        EventUri uri = EventUri.fromString("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2");
+        EventUri uri = EventUri.parse("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2");
         assertThat(uri.getKind().getValue(), is(1));
         assertThat(uri.getPublicKeyHex(), is("82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2"));
         assertThat(uri.getIdentifier().isPresent(), is(false));
@@ -41,7 +41,7 @@ class EventUriTest {
 
     @Test
     void createFromString2() {
-        EventUri uri = EventUri.fromString("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:");
+        EventUri uri = EventUri.parse("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:");
         assertThat(uri.getKind().getValue(), is(1));
         assertThat(uri.getPublicKeyHex(), is("82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2"));
         assertThat(uri.getIdentifier().orElseThrow(), is(""));
@@ -49,7 +49,7 @@ class EventUriTest {
 
     @Test
     void createFromString3() {
-        EventUri uri = EventUri.fromString("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:test");
+        EventUri uri = EventUri.parse("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:test");
         assertThat(uri.getKind().getValue(), is(1));
         assertThat(uri.getPublicKeyHex(), is("82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2"));
         assertThat(uri.getIdentifier().orElseThrow(), is("test"));
@@ -57,7 +57,7 @@ class EventUriTest {
 
     @Test
     void createFromString4() {
-        EventUri uri = EventUri.fromString("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:::::: test :::::");
+        EventUri uri = EventUri.parse("1:82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2:::::: test :::::");
         assertThat(uri.getKind().getValue(), is(1));
         assertThat(uri.getPublicKeyHex(), is("82341f882b6eabcd2ba7f1ef90aad961cf074af15b9ef44a09f9d2a8fbfbe6a2"));
         assertThat(uri.getIdentifier().orElseThrow(), is("::::: test :::::"));